# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Instay backend API documentation",
        "title": "Instay API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/service-types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lấy tất cả loại dịch vụ cho admin xem",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Get Service Types for Admin",
                "responses": {
                    "200": {
                        "description": "Lấy tất cả service type thành công",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.ServiceTypeResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Invalid Information",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Tạo một loại dịch vụ mới",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Create Service Type",
                "parameters": [
                    {
                        "description": "Thông tin loại dịch vụ mới",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateServiceTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tạo service type thành công",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request (validation error)",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Department không tìm thấy",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (service type name đã tồn tại)",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/service-types/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Xoá một loại dịch vụ bằng ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Delete Service Type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xoá service type thành công",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request (ID không hợp lệ)",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Service type không tìm thấy",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (không thể xoá bản ghi được bảo vệ)",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cập nhật thông tin của một loại dịch vụ bằng ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Update Service Type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin cần cập nhật",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateServiceTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cập nhật service type thành công",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request (validation, ID, logic error)",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Department Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (service type name đã tồn tại)",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/service-types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lấy tất cả loại dịch vụ cho khách xem",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Get Service Types for Guest",
                "responses": {
                    "200": {
                        "description": "Lấy tất cả service type thành công",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.SimpleServiceTypeResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "types.BasicUserResponse": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "types.CreateServiceTypeRequest": {
            "type": "object",
            "required": [
                "department_id",
                "name"
            ],
            "properties": {
                "department_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
...
        "types.ServiceTypeResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "$ref": "#/definitions/types.BasicUserResponse"
                },
                "department": {
                    "$ref": "#/definitions/types.SimpleDepartmentResponse"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "service_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "$ref": "#/definitions/types.BasicUserResponse"
                }
            }
        },
        "types.SimpleDepartmentResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "types.SimpleServiceTypeResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateServiceTypeRequest": {
            "type": "object",
            "properties": {
                "department_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
...
    }
}